# swagger/swagger.json
{
	"components": {
		"schemas": {
...
			"malak.Dashboard": {
				"properties": {
					"chart_count": {
						"type": "integer"
					},
					"created_at": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"reference": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"workspace_id": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"malak.DashboardChart": {
				"properties": {
					"chart": {
						"$ref": "#/components/schemas/malak.IntegrationChart"
					},
					"chart_id": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"dashboard_id": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"reference": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"workspace_id": {
						"type": "string"
					},
					"workspace_integration_id": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"malak.DashboardChartPosition": {
				"properties": {
					"chart_id": {
						"type": "string"
					},
					"dashboard_id": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"order_index": {
						"type": "integer"
					}
				},
				"type": "object"
			},
...
			"malak.IntegrationChart": {
				"properties": {
					"chart_type": {
						"$ref": "#/components/schemas/malak.IntegrationChartType"
					},
					"created_at": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"internal_name": {
						"$ref": "#/components/schemas/malak.IntegrationChartInternalNameType"
					},
					"metadata": {
						"$ref": "#/components/schemas/malak.IntegrationChartMetadata"
					},
					"reference": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"user_facing_name": {
						"type": "string"
					},
					"workspace_id": {
						"type": "string"
					},
					"workspace_integration_id": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"malak.IntegrationChartInternalNameType": {
				"enum": [
					"mercury_account",
					"mercury_account_transaction",
					"brex_account",
					"brex_account_transaction"
				],
				"type": "string",
				"x-enum-varnames": [
					"IntegrationChartInternalNameTypeMercuryAccount",
					"IntegrationChartInternalNameTypeMercuryAccountTransaction",
					"IntegrationChartInternalNameTypeBrexAccount",
					"IntegrationChartInternalNameTypeBrexAccountTransaction"
				]
			},
			"malak.IntegrationChartMetadata": {
				"properties": {
					"provider_id": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"malak.IntegrationChartType": {
				"enum": [
					"bar",
					"pie"
				],
				"type": "string",
				"x-enum-varnames": [
					"IntegrationChartTypeBar",
					"IntegrationChartTypePie"
				]
			},
			"malak.IntegrationDataPoint": {
				"properties": {
					"created_at": {
						"type": "string"
					},
					"data_point_type": {
						"$ref": "#/components/schemas/malak.IntegrationDataPointType"
					},
					"id": {
						"type": "string"
					},
					"integration_chart_id": {
						"type": "string"
					},
					"metadata": {
						"$ref": "#/components/schemas/malak.IntegrationDataPointMetadata"
					},
					"point_name": {
						"type": "string"
					},
					"point_value": {
						"type": "integer"
					},
					"reference": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"workspace_id": {
						"type": "string"
					},
					"workspace_integration_id": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"malak.IntegrationDataPointMetadata": {
				"type": "object"
			},
			"malak.IntegrationDataPointType": {
				"enum": [
					"currency",
					"others"
				],
				"type": "string",
				"x-enum-varnames": [
					"IntegrationDataPointTypeCurrency",
					"IntegrationDataPointTypeOthers"
				]
			},
...
			"server.APIStatus": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object"
			},
			"server.addChartToDashboardRequest": {
				"properties": {
					"chart_reference": {
						"type": "string"
					}
				},
				"required": [
					"chart_reference"
				],
				"type": "object"
			},
...
			"server.createDashboardRequest": {
				"properties": {
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"description",
					"title"
				],
				"type": "object"
			},
...
			"server.fetchDashboardResponse": {
				"properties": {
					"dashboard": {
						"$ref": "#/components/schemas/malak.Dashboard"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"dashboard",
					"message"
				],
				"type": "object"
			},
...
			"server.listChartDataPointsResponse": {
				"properties": {
					"data_points": {
						"items": {
							"$ref": "#/components/schemas/malak.IntegrationDataPoint"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data_points",
					"message"
				],
				"type": "object"
			},
...
			"server.listDashboardChartsResponse": {
				"properties": {
					"charts": {
						"items": {
							"$ref": "#/components/schemas/malak.DashboardChart"
						},
						"type": "array"
					},
					"dashboard": {
						"$ref": "#/components/schemas/malak.Dashboard"
					},
					"message": {
						"type": "string"
					},
					"positions": {
						"items": {
							"$ref": "#/components/schemas/malak.DashboardChartPosition"
						},
						"type": "array"
					}
				},
				"required": [
					"charts",
					"dashboard",
					"message",
					"positions"
				],
				"type": "object"
			},
			"server.listDashboardResponse": {
				"properties": {
					"dashboards": {
						"items": {
							"$ref": "#/components/schemas/malak.Dashboard"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"meta": {
						"$ref": "#/components/schemas/server.meta"
					}
				},
				"required": [
					"dashboards",
					"message",
					"meta"
				],
				"type": "object"
			},
			"server.listIntegrationChartsResponse": {
				"properties": {
					"charts": {
						"items": {
							"$ref": "#/components/schemas/malak.IntegrationChart"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"charts",
					"message"
				],
				"type": "object"
			},
...
			"server.meta": {
				"properties": {
					"paging": {
						"$ref": "#/components/schemas/server.pagingInfo"
					}
				},
				"required": [
					"paging"
				],
				"type": "object"
			},
			"server.pagingInfo": {
				"properties": {
					"page": {
						"type": "integer"
					},
					"per_page": {
						"type": "integer"
					},
					"total": {
						"type": "integer"
					}
				},
				"required": [
					"page",
					"per_page",
					"total"
				],
				"type": "object"
			},
...
			"server.updateDashboardPositionsRequest": {
				"properties": {
					"positions": {
						"items": {
							"properties": {
								"chart_id": {
									"type": "string"
								},
								"index": {
									"type": "integer"
								}
							},
							"required": [
								"chart_id",
								"index"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"positions"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"in": "header",
				"name": "Authorization",
				"type": "apiKey"
			}
		}
	},
	"info": {
		"contact": {
			"email": "lanre@ayinke.ventures",
			"name": "Ayinke Ventures"
		},
		"title": "Malak's API documentation",
		"version": "0.1.0"
	},
	"openapi": "3.0.3",
	"paths": {
...
		"/dashboards": {
			"get": {
				"description": "List dashboards",
				"parameters": [
					{
						"description": "Page to query data from. Defaults to 1",
						"in": "query",
						"name": "page",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "Number to items to return. Defaults to 10 items",
						"in": "query",
						"name": "per_page",
						"schema": {
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.listDashboardResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"dashboards"
				]
			},
			"post": {
				"description": "create a new dashboard",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/server.createDashboardRequest"
							}
						}
					},
					"description": "dashboard request body",
					"required": true,
					"x-originalParamName": "message"
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.fetchDashboardResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"dashboards"
				]
			}
		},
		"/dashboards/charts": {
			"get": {
				"description": "List charts",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.listIntegrationChartsResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"dashboards"
				]
			}
		},
		"/dashboards/charts/{reference}": {
			"get": {
				"description": "fetch charting data",
				"parameters": [
					{
						"description": "chart unique reference.. e.g integration_chart_km31C.e6xV",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.listChartDataPointsResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"dashboards"
				]
			}
		},
		"/dashboards/{reference}": {
			"get": {
				"description": "fetch dashboard",
				"parameters": [
					{
						"description": "dashboard unique reference.. e.g dashboard_",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.listDashboardChartsResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"dashboards"
				]
			}
		},
		"/dashboards/{reference}/charts": {
			"put": {
				"description": "add a chart to a dashboard",
				"parameters": [
					{
						"description": "dashboard unique reference.. e.g dashboard_",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/server.addChartToDashboardRequest"
							}
						}
					},
					"description": "dashboard request chart data",
					"required": true,
					"x-originalParamName": "message"
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"dashboards"
				]
			}
		},
		"/dashboards/{reference}/positions": {
			"post": {
				"description": "update dashboard chart positioning",
				"parameters": [
					{
						"description": "dashboard unique reference.. e.g dashboard_22",
						"in": "path",
						"name": "reference",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/server.updateDashboardPositionsRequest"
							}
						}
					},
					"description": "dashboard chart positions",
					"required": true,
					"x-originalParamName": "message"
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/server.APIStatus"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"dashboards"
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:5300/v1"
		}
	]
}